+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## [Unreleased]
 
-- add werror, debug options.
+- add werror, info, debug options.
 - modify parse front matter logic.
 
 ## [0.4.0] - 2022-04-09